move RingBuffer from API to SDK dts
